crypto: ccp: remove multi-level pointers processing for vpsp
hygon inclusion
category: feature

---------------------------

The data of the CSV VM is encrypted, so the ciphertext
data of the CSV VM cannot be copied to the host kernel
space for processing.

This patch removes the old copy-based processing.
